+++ main_page.py
@@ -1,5 +1,4 @@
 import streamlit as st
-import one_gene_search
 
 def create_header():
     st.image('images/logo.png', width=100)
